+++ synapse/common.py
@@ -912,6 +912,10 @@ def deprecated(name, curv='2.x', eolv='3.0.0'):
     mesg = f'"{name}" is deprecated in {curv} and will be removed in {eolv}'
     warnings.warn(mesg, DeprecationWarning)
 
+splicedepr = '2023-10-01'
+def deprdate(name, date):
+    warnings.warn(f'{name} is deprecated and will be removed on {date}.')
+
 def reqjsonsafe(item):
     '''
     Returns None if item is json serializable, otherwise raises an exception.